week7(practice)/Day5/requirement.txt
pymupdf
pytesseract
pillow 
sha256
tqdm
fastapi 
uvicorn
redis
streamlt 
request 
